+++ src/uwtools/tests/config/formats/test_yaml.py
@@ -1,0 +1,146 @@
+"""
+Tests for uwtools.config.formats.yaml module.
+"""
+
+import datetime
+import filecmp
+import logging
+import sys
+from io import StringIO
+from unittest.mock import patch
+
+import yaml
+from pytest import raises
+
+from uwtools import exceptions
+from uwtools.config import support
+from uwtools.config.formats.yaml import YAMLConfig
+from uwtools.exceptions import UWConfigError
+from uwtools.logging import log
+from uwtools.tests.support import fixture_path, logged
+from uwtools.utils.file import _stdinproxy
+
+# Tests
+
+
+def test_composite_types():
+    """
+    Test that YAML load and dump work with a YAML file that has multiple data structures and levels.
+    """
+    cfgobj = YAMLConfig(fixture_path("result4.yaml"))
+
+    assert cfgobj["step_cycle"] == "PT6H"
+    assert isinstance(cfgobj["init_cycle"], datetime.datetime)
+
+    generic_repos = cfgobj["generic_repos"]
+    assert isinstance(generic_repos, list)
+    assert isinstance(generic_repos[0], dict)
+    assert generic_repos[0]["branch"] == "develop"
+
+    models = cfgobj["models"]
+    assert models[0]["config"]["vertical_resolution"] == 64
+
+
+def test_include_files():
+    """
+    Test that including files via the include constructor works as expected.
+    """
+    cfgobj = YAMLConfig(fixture_path("include_files.yaml"))
+
+    # 1-file include tests.
+
+    assert cfgobj["salad"]["fruit"] == "papaya"
+    assert cfgobj["salad"]["how_many"] == 17
+    assert len(cfgobj["salad"]) == 4
+
+    # 2-file test, checking that values provided by the first file are replaced
+    # by values from the second file. There should be 7 items under two_files.
+
+    assert cfgobj["two_files"]["fruit"] == "papaya"
+    assert cfgobj["two_files"]["vegetable"] == "peas"
+    assert len(cfgobj["two_files"]) == 7
+
+    # 2-file test, but with included files reversed.
+
+    assert cfgobj["reverse_files"]["vegetable"] == "eggplant"
+
+
+def test_simple(tmp_path):
+    """
+    Test that YAML load, update, and dump work with a basic YAML file.
+    """
+    infile = fixture_path("simple2.yaml")
+    outfile = tmp_path / "outfile.yml"
+    cfgobj = YAMLConfig(infile)
+    expected = {
+        "account": "user_account",
+        "extra_stuff": 12345,
+        "jobname": "abcd",
+        "nodes": 1,
+        "queue": "bos",
+        "scheduler": "slurm",
+        "tasks_per_node": 4,
+        "walltime": "00:01:00",
+    }
+    assert cfgobj == expected
+    cfgobj.dump(outfile)
+    assert filecmp.cmp(infile, outfile)
+    cfgobj.update({"nodes": 12})
+    expected["nodes"] = 12
+    assert cfgobj == expected
+
+
+def test_constructor_error_no_quotes(tmp_path):
+    # Test that Jinja2 template without quotes raises UWConfigError.
+
+    tmpfile = tmp_path / "test.yaml"
+    with tmpfile.open("w", encoding="utf-8") as f:
+        f.write(
+            """
+foo: {{ bar }}
+bar: 2
+"""
+        )
+    with raises(exceptions.UWConfigError) as e:
+        YAMLConfig(tmpfile)
+    assert "value is enclosed in quotes" in str(e.value)
+
+
+def test_constructor_error_unregistered_constructor(tmp_path):
+    # Test that unregistered constructor raises UWConfigError.
+
+    tmpfile = tmp_path / "test.yaml"
+    with tmpfile.open("w", encoding="utf-8") as f:
+        f.write("foo: !not_a_constructor bar")
+    with raises(exceptions.UWConfigError) as e:
+        YAMLConfig(tmpfile)
+    assert "constructor: '!not_a_constructor'" in str(e.value)
+    assert "Define the constructor before proceeding" in str(e.value)
+
+
+def test_YAMLConfig__load_unexpected_error(tmp_path):
+    cfgfile = tmp_path / "cfg.yaml"
+    with open(cfgfile, "w", encoding="utf-8") as f:
+        print("{n: 88}", file=f)
+    with patch.object(yaml, "load") as load:
+        msg = "Unexpected error"
+        load.side_effect = yaml.constructor.ConstructorError(note=msg)
+        with raises(UWConfigError) as e:
+            YAMLConfig(config_file=cfgfile)
+        assert msg in str(e.value)
+
+
+def test_YAMLConfig__load_paths_failure_stdin_plus_relpath(caplog):
+    # Instantiate a YAMLConfig with no input file, triggering a read from stdin. Patch stdin to
+    # provide YAML with an include directive specifying a relative path. Since a relative path
+    # is meaningless relative to stdin, assert that an appropriate error is logged and exception
+    # raised.
+    log.setLevel(logging.INFO)
+    _stdinproxy.cache_clear()
+    relpath = "../bar/baz.yaml"
+    with patch.object(sys, "stdin", new=StringIO(f"foo: {support.INCLUDE_TAG} [{relpath}]")):
+        with raises(UWConfigError) as e:
+            YAMLConfig()
+    msg = f"Reading from stdin, a relative path was encountered: {relpath}"
+    assert msg in str(e.value)
+    assert logged(caplog, msg)
